bolt-reporting/requirements.txt
Jinja2==3.1.2
matplotlib==3.5.2
numpy==1.22.3
pandas==1.4.2
python-dateutil==2.8.2
requests==2.27.1
weasyprint==52.5
XlsxWriter==3.0.3
google-cloud-storage==2.4.0
google==3.0.0
